fix: compile-error with older avr-gcc in cpucontext due to memory barrier
Older gccs seem to have a problem with empty asm statements.
(This is statements with no asm opcode instruction)
Since the usual "_MemoryBarrier" from avr-libc macro uses such an empty
statement, compilation fails.

The fix overrides the avr-libc definition of the barrier in case
of older version or in case "CPUCONTEXT_REPLACEMEMBARRIER" is defined.

Signed-off-by: Stephan Baerwolf <stephan@matrixstorm.com>